Move files to base module lib
